HW1. corrections3

+++ src/main/java/ru/javawebinar/topjava/model/Meal.java
@@ -5,19 +5,19 @@ import java.time.LocalDateTime;
 import java.time.LocalTime;
 
 public class Meal {
-    private long id;
+    private Integer id;
     private final LocalDateTime dateTime;
     private final String description;
     private final int calories;
 
-    public Meal(long id, LocalDateTime dateTime, String description, int calories) {
+    public Meal(Integer id, LocalDateTime dateTime, String description, int calories) {
         this.id = id;
         this.dateTime = dateTime;
         this.description = description;
         this.calories = calories;
     }
 
-    public long getId() {
+    public Integer getId() {
         return id;
     }
 
@@ -41,7 +41,7 @@ public class Meal {
         return dateTime.toLocalTime();
     }
 
-    public void setId(long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 }
